MFC r281606:
In the version of gcc in the FreeBSD tree, this modification was made to
the compiler in svn r242182:

A similar change was done to clang in the FreeBSD tree in svn r218893:

However, for external gcc toolchains, this patch is not in the compiler's header
file.

This patch to FreeBSD's aesni code allows compilation with an external
gcc toolchain.

Differential Revision: https://reviews.freebsd.org/D2285
Reviewed by:  jmg, dim
Approved by:  dim

TAG:	IPSEC-HEAD
Issue:	#4841