Do not build lint(1) by default on stable-11, add WITH_LINT to enable building it.
lint(1) is required on the host build system to build lint(1) libraries, and
it is no longer available on FreeBSD 12. This prevents using a 12 or later
host to build an 11 or earlier system, which causes problems for building
jails and using poudriere. The problem could be fixed by treating lint as a
bootstrap tool when building on 12+, but that just adds complexity and build
time to build a broken tool that virtually nobody uses anymore.

This is a direct commit to 11-stable because lint no longer exists in later
branches.

PR:		223892
Relnotes:	yes
Differential Revision:	https://reviews.freebsd.org/D13799

(cherry picked from commit 5fb1dbc1862d5ddd058d22fe18063e6c71aeb7bc)