Detect and fix case in wg_init() where sockets weren't cleaned up.
Based on inspection, it appears possible for wg_init() to be called
without a corresponding wg_stop(). If this happens, the UDP sockets
could be left dangling, resulting in a crash on packet input.

Printf an XXX to the console and clean up the socket if this is
detected.

Ticket: #5186