Return immediately in wg_init if in DETACH'd state.
The iflib ioctl	code will force	a re-init in certain conditions,
even if	the interface is in the	process	of being destroyed. This
would result in	a new listening	socket being allocated,	which would
be lost	when the softc was finally destroyed.

Next packet arrival would match	this listening socket, resulting in
an upcall using	a non-existent softc and a panic.

For for	the time being by immediately returning	in wg_init if the
interface is being detached. However, more investigation is needed to
see if the 12-stable iflib common code should be updated to handle
this case.

Ticket: #5186