Fix issue with replying	to INITIATION messages in server mode.
Set the remote peer address/port from the incoming INITIATION message
once decryption	has succeeded. This allows the reply to	be sent
if a remote peer address/port wasn't configured, which is standard
practice in server mode.

Tested with an OpenBSD client and a server-style config. Additional
testing	with changing the host address of the client and verifying
continued tunnel connectivity.

ticket: #5509