Fix the autoboot timer on aarch64.
Using higher delay times here, causes the loader.time() function to lag behind,
which show up as a long delay in the countdown timer at menu.

It is not clear what causes the delay, as the implementation goes back to EFI
support in u-boot, but once you reduce the busy loop times it gets happy.

Task:	#5334