safexcel: Fix the SHA-HMAC digest computation when AAD is present
The driver would fail to include the AAD in the input stream, resulting
in incorrect digests for requests combining SHA-HMAC with AES-CBC or
-CTR.  Ensure that the AAD is included in the processor's input stream,
and fix the corresponding instruction sequence to include the AAD as
input to the digest computation.

This is a direct commit to stable/12 since the bug was introduced while
merging there and is not present in later branches.

(cherry picked from commit d74d1284386aa37e6ec8ffa7739c51b3fa54fcb3)