Use autonegotiation to set 10/100 link speeds.
Earlier	versions of this driver	had PHY	drivers	that would
accept forced 10/100 link speeds. However, these have been removed,
and the	Linux driver only allows autonegotation	for these speeds.
Forcing	the link speed would result in a call to a null routine,
resulting in the link being set	to auto.

While here, remove the case for 1000 LX/SX fiber: the igc is copper-only.

Tested with 10/100 half/full duplex and	1000-baseT to an unmanaged autoneg
switch,	and also 2 igc ports back-back with one	set to auto and	the other
with all settings of 10/100 half/full and 1000.

iperf3 was used	to verify performance at all link settings.

Ticket:	#6069