ng_l2tp: workaround a race between ng_l2tp_seq_rack_timeout and resetting
ng_uncallout while the callout is pending fails to cancel it. At the
same time ng_l2tp_seq_rack_timeout is blocked waiting on the seq->mtx
lock. Once the lock is released the func proceeds to check for
cancellation which we know failed and then incorrectly assuming there is
work to do.

Work around for now by NULL checking the queue.