pf: revert pf_route_to_daemon_friendly.diff
The current version of this patch is broken. It passes an offset of
'ntohs(ip->ip_off)' to PACKET_UNDO_NAT(). The offset is supposed to point to
the start of the IP payload (i.e. the TCP/UDP/ICMP/... header). However, it
uses the fragment offset for this rather than the IP header length.
Also, it fails to account for the flag bits on ip_off, so even if it wanted
the fragment offset this would still be wrong.

This implies that the patch isn't required for current use cases, so rather
than carry a local difff we revert is, and more closely align with upstream.

Redmine:	7072
(cherry picked from commit 325e8fc67a3d0e4ec0947e6e0de1dfdc79400e30)