Separate out build src.conf from install src.conf
This will let a poudriere jail be built that has things not present in
the installed image.  The reason for this is some ports check for silly
things, even when they do not use them.  Like arpwatch ./configure
requires a sendmail binary, even when it is not needed for build and a
package will provide it later.