efi: loader: Inline copy_finish function in amd64 trampoline
Instead of calling the efi_copy_finish function from amd64_tramp,
include the copy instructions in the trampoline code itself.
This avoids boot hangs and restarts in the cases when
the efi_copy_finish code happens to be located at a memory
location that is overwritten during the copy process.

This is a direct commit to stable/12 since no-copy loader mode requires
the kernel update not suitable for the branch.

PR:	209821