pfctl: remove remnants of PF_USER_INFO
This no longer serves any purpose now that 'ridentifier' is upstream and
usef by pfSense. This brings us closer to upstream, and returns some
output to the upstream default as well.

Specifically, it removes the '(0)' section in rules such as:
@4(0) block drop in log inet all label "Default deny rule IPv4" ridentifier 1000000103

This is where the trackerid (now ridentifier) used to be printed, but
it's currently always zero.

Redmine:	12868


(cherry picked from commit dd728141710ecc42d8f1427f7b9454742a811969)