epoch: fix assertion
If we call epoch_exit_preempt() on an epoch that's not net_epoch_preempt
we can't assert td->td_net_epoch.
Move the assertion under the 'epoch == net_epoch_preempt' conditional.

No functional change for pfSense, as we don't build with INVARIANTS
(unless debugging, in which case this panicked during boot).