pf: allow export of creator ids
It is sometimes useful to obtain a list of creator ids, but it is also
inefficient to export the full state list to summarise this information
in userspace. Assemble the list in the kernel, removing the need to
export all of the states first.

Sponsored by:	Rubicon Communications, LLC ("Netgate")