ixgbe: reduce waits for I2C
If we've not identified the SFP type do not retry. That is, try each I2C
read operation only once.

Sadly there do not appear to be any reliably interrupts to tell us that an
SFP has been (un)plugged, so we cannot rely on anything to just avoid the
I2C timeout entirely.

Redmine:	https://redmine.pfsense.org/issues/13911