Always update uefi boot entry in bootconfig
Summary:
Change 40923b0c81cc introduced a conditional check on BSDINSTALL_CONFIGCURRENT
to control whether or not to run efibootmgr to add and activate and entry for
the loader installed to the destination ESP. This prevents diskless scripted
installers from installing a bootable system where the EFI boot entry does not
already exist.

This change removes the conditional on BSDINSTALL_CONFIGCURRENT, so that on a
diskless installer the boot entry is added and activated as expected.

Reviewers: nwhitehorn, dteske

Subscribers: imp, jrtc27

Differential Revision: https://reviews.freebsd.org/D42535

(cherry picked from commit daa9ed55c40aafb509e99eb0aa42c85c65780850)