lib/libpfctl: return to a private library
Stop building the base system libpfctl as a public library. We now have a
ports version of libpfctl, which ports can build against. This brings us
closer to upstream, and also means it's easier to fix issues in libpfctl
(because a ports update is easier than a base system update).

Revert "Make libpfctl a public library."
This reverts commit 9e31e6ad1c36e55a26a48e046a9a424a48ca5330.

Revert "Add libnv to the libraries list in the libpfctl build."
This reverts commit a34aa14149cc3f4a5c53e11ad141f77f724cac81.