llvm: default undefined_version on
This is equivalent to building everything with -Wl,-undefined_version, and
ensures builds do not fail when a linker script mentions undefined functions.
That in turn fixes a number of ports build issues with llvm 17.

The expectation is that upstream (i.e. freebsd ports maintainers) will
gradually resolve these problems, but in the mean time we'll work around them.

This patch is to be reverted as soon the ports we care about are fixed.