netinet6: work around #14431
For unclear reasons we can end up with route entries that use interfaces
that do not have the IPv6 protocol family. This is assumed to be a race
condition, when the relevant interface is being destroyed.

While that's being debugged we can hopefully avoid the panic by
explicitly checking that the returned interface does have an IPv6
address family before we return it as a result of the route lookup.

Redmine:	https://redmine.pfsense.org/issues/14431