pfsync: Fix offset calculation
Even though message version is automatically recognized and the top of
the struct is identical for different versions, when iterating over
multiple messages proper message length must be used. That's the length
of an union member for given version, not of the union itself.

Reviewed by:	kp
Differential Revision:	https://reviews.freebsd.org/D43862

(cherry picked from commit 50edc630719827b6c58dd515328997fd196b1d78)