bxe: Add 2.5G SGMII support (SFP GPON ONT)
This is an old well-known patch (from 2018) originating from here:
https://www.dslreports.com/forum/r32230041-Internet-Bypassing-the-HH3K-up-to-2-5Gbps-using-a-BCM57810S-NIC

So this patch was adapted to the current head from the patch posted here:
https://www.dslreports.com/forum/r32230853-

It allows to reach 2.5Gbps with GPON SFP on QLogic NetXtreme II BCM57810.

This patch is used by lot of pfSense/OPNSense and FreeBSD user on their home router.
This will ease life of all home fiber users.

Differential Revision:	https://reviews.freebsd.org/D36508