Add a sysctl to control selection of CARP addresses for source ports
Nexus binds sockets for netgard tunnels to INADDR_ANY for connection to a
controller. This has the downside that in the rare instance that a CARP address
precedes a unique address in the interface address list, the shared CARP address
can be chosen as the source port leading to a scenario where the two HA systems
are not reliably distinguished by address. To resolve this, it is desirable to
prevent the selection of CARP addresses when selecting the source address
e.g. from fib4_lookup().

Add net.route.allow_carp_src sysctl to control the behavior of ifaof_ifpforaddr()
to prefer non-CARP addresses for source addresses on sockets.